drug_regimen_sunburst

This function allows the user to visualize the complete treatment course for selected diagnoses.

Usage

drug_regimen_sunburst(data_synapse, data_cohort, max_n_regimens = NULL)

Arguments

data_synapse

The list returned from the `pull_data_synapse()` function call

data_cohort

The list returned from the `create_analytic_cohort()` function call

max_n_regimens

The number of regimens displayed in the sunburst plot

Value

Returns data frame `treatment_history` and interactive plot `sunburst_plot`

Examples

if(genieBPC:::check_synapse_login() == FALSE){
nsclc_data <- pull_data_synapse("NSCLC", version = "2.1-consortium")
nsclc_cohort <- create_analytic_cohort(cohort = "NSCLC",
data_synapse = nsclc_data, stage = "Stage IV")
ex1 <- drug_regimen_sunburst(data_synapse = nsclc_data,
data_cohort = nsclc_cohort,
max_n_regimens = 3)
}
#> Welcome, Michael Curry!
#> 
#> [1] "Great job! You are logged into Synapse and\n        can access the GENIE BPC data."
#> Welcome, Michael Curry!
#>