fetch_samples

Subset cancer panel test data to patients in the cohort of interest

fetch_samples(cohort, data_synapse, df_record_ids)

Arguments

cohort
GENIE BPC Project cancer. Must be one of "NSCLC" or "CRC".

data_synapse
the list object outputted by the `pull_data_synapse()` function.

df_record_ids
output object of the create_cohort function.

Value

returns the cohort object list inputted with an additional dataset named "samples_data".

Author

Axel Martin #' # Example 1 ---------------------------------- # Create a cohort of all patients with stage IV NSCLC of histology adenocarcinoma out <- create_cohort(cohort = "NSCLC", stage_dx = c("Stage IV"), ca_hist_adeno_squamous = "Adenocarcinoma") samples_data <- fetch_samples(cohort = "NSCLC", data_synapse = out$cohort_ca_dx) # Example 2 ---------------------------------- # Create a cohort of all NSCLC patients who received Cisplatin, Pemetrexed Disodium or Cisplatin, Etoposide as their first drug regimen out <- create_cohort(cohort = "NSCLC", regimen_drugs = c("Cisplatin, Pemetrexed Disodium", "Cisplatin, Etoposide"), regimen_order = 1, regimen_order_type = "within regimen") samples_data <- fetch_samples(cohort = "NSCLC", data_synapse = out$cohort_ca_dx)