genieBPC: Project GENIE BioPharma Collaborative Data Processing Pipeline

The American Association Research (AACR) Project Genomics Evidence Neoplasia Information Exchange (GENIE) BioPharma Collaborative represents a multi-year, multi-institution effort to build a pan-cancer repository of linked clinico-genomic data. The genomic and clinical data are provided in multiple releases (separate releases for each cancer cohort with updates following data corrections), which are stored on the data sharing platform 'Synapse' https://www.synapse.org/. The 'genieBPC' package provides a seamless way to obtain the data corresponding to each release from 'Synapse' and to prepare datasets for analysis.

See also

Author

Maintainer: Jessica A. Lavery laveryj@mskcc.org (ORCID)

Authors:

Michael A. Curry (ORCID)

Samantha Brown (ORCID)

Karissa Whiting (ORCID)

Axel Martin

Other contributors:

Daniel D. Sjoberg danield.sjoberg@gmail.com (ORCID) [contributor]