select_unique_ngs

Get a unique next generation sequencing sample for each patient for analysis following several user define criterions.

select_unique_ngs(
  data_cohort,
  oncotree_code = NULL,
  sample_type = NULL,
  min_max_time = NULL
)

Arguments

data_cohort

output object of the create_analytic_cohort function.

oncotree_code

character vector specifying which sample OncoTree codes to keep. See "cpt_oncotree_code" column of data_cohort argument above to get options.

sample_type

character specifying which type of genomic sample to prioritize, options are "Primary", "Local" and "Metastasis". Default is to not select a NGS sample based on the sample type.

min_max_time

character specifying if the first or last genomic sample recorded should be kept. Options are "min" (first) and "max" (last).

Value

returns the sample object list inputted with an additional dataset named "samples_data".

Examples

# Example 1 ----------------------------------
# Create a cohort of all patients with stage IV NSCLC of histology
# adenocarcinoma
# out <- create_analytic_cohort(cohort = "NSCLC",
#      stage_dx = c("Stage IV"),
#      ca_hist_adeno_squamous = "Adenocarcinoma")
# select_unique_ngs <- select_unique_ngs(data_cohort = out)
# Example 2 ----------------------------------
# Create a cohort of all NSCLC patients who received Cisplatin,
# Pemetrexed Disodium or Cisplatin,
# Etoposide as their first drug regimen
# out <- create_analytic_cohort(cohort = "NSCLC",
#     regimen_drugs = c("Cisplatin, Pemetrexed Disodium",
# "Cisplatin, Etoposide"),
#     regimen_order = 1,
#     regimen_order_type = "within regimen")
# select_unique_ngs <- select_unique_ngs(data_cohort = out,
# oncotree_code = "LUAD", sample_type = "Metastasis",min_max_time = "max")