create_analytic_cohort

This function allows the user to create a cohort from the GENIE BPC data based on cancer diagnosis information such as cancer cohort, treating institution, histology, and stage at diagnosis, as well as cancer-directed regimen information including regimen name and regimen order. This function returns three datasets:

`cohort_ca_dx` will contain cancer diagnosis information for patients matching the specified criteria. This dataset is structured as one record per patient per associated cancer diagnosis.

`cohort_ca_drugs` will return the drug-regimen information for patients matching the specified criteria. This dataset is structured as one record per patient, per regimen and associated cancer diagnosis. For example, a drug regimen associated with treating two cancer diagnoses will have two records in this dataset.

`cohort_ngs` is the next generation sequencing report information for patients matching the specified criteria. This dataset is structured as one record per next generation sequencing report, per associated diagnosis. If it is unknown which cancer sequence is associated with the next generation sequencing report, it will have two records in this dataset. This dataset contains the key to merge the clinical and genomic data.

The function inputs `cohort`, `institution`, `stage_dx`, and `regimen_drugs` correspond to the variable names in the GENIE BPC Analytic Data Guide, available on Synapse.

Usage

create_analytic_cohort(
  cohort,
  data_synapse,
  index_ca_seq = 1,
  institution,
  stage_dx,
  histology,
  regimen_drugs,
  regimen_type = "Exact",
  regimen_order,
  regimen_order_type,
  return_summary = FALSE
)

Arguments

cohort

GENIE BPC Project cancer. Must be one of "NSCLC" (non-small cell lung cancer) or "CRC" (colorectal cancer). Future cohorts will include "BrCa" (breast cancer), "PANC" (pancreatic cancer), "Prostate" (prostate cancer).

data_synapse

The list returned from pull_data_synapse().

index_ca_seq

Index cancer sequence. Default is 1, indicating the patient's first index cancer. The index cancer is also referred to as the BPC Project cancer in the GENIE BPC Analytic Data Guide; this is the cancer that met the eligibility criteria for the project and was selected at random for PRISSMM phenomic data curation. Specifying multiple index cancer sequences, e.g. index_ca_seq = c(1, 2) will return index cancers to patients with 1 index cancer and will return the first AND second index cancers to patients with multiple.

institution

GENIE BPC participating institution. Must be one of "DFCI", "MSK", "UHN", or "VICC" for NSCLC cohorts; must be one of "DFCI", "MSK", "VICC" for CRC. Default selection is all institutions. This parameter corresponds to the variable `institution` in the Analytic Data Guide.

stage_dx

Stage at diagnosis. Must be one of "Stage I", "Stage II", "Stage III", "Stage I-III NOS", "Stage IV". Default selection is all stages. This parameter corresponds to the variable `stage_dx` in the Analytic Data Guide.

histology

Cancer histology. For all cancer cohorts except for BrCa (breast cancer), this parameter corresponds to the variable `ca_hist_adeno_squamous` and must be one of "Adenocarcinoma", "Squamous cell", "Sarcoma", "Small cell carcinoma", "Carcinoma", "Other histologies/mixed tumor". For BrCa, this parameter corresponds to the variable `ca_hist_brca` and must be one of "Invasive lobular carcinoma", "Invasive ductal carcinoma", "Other histology". Default selection is all histologies.

regimen_drugs

Vector with names of drugs in cancer-directed regimen, separated by a comma. For example, to specify a regimen consisting of Carboplatin and Pemetrexed, specify regimen_drugs = "Carboplatin, Pemetrexed". Acceptable values are found in the `drug_regimen_list` dataset provided with this package. This parameter corresponds to the variable `regimen_drugs` in the Analytic Data Guide.

regimen_type

Indicates whether the regimen(s) specified in `regimen_drugs` indicates the exact regimen to return, or if regimens containing the drugs listed in `regimen_drugs` should be returned. Must be one of "Exact" or "Containing". The default is "Exact".

regimen_order

Order of cancer-directed regimen. If multiple drugs are specified, `regimen_order` indicates the regimen order for all drugs; different values of `regimen_order` cannot be specified for different drug regimens. If multiple values are specified, e.g. c(1, 2), then drug regimens that met either order criteria are returned.

regimen_order_type

Specifies whether the `regimen_order` parameter refers to the order of receipt of the drug regimen within the cancer diagnosis (across all other drug regimens; "within cancer") or the order of receipt of the drug regimen within the times that that drug regimen was administered (e.g. the first time carboplatin pemetrexed was received, out of all times that the patient received carboplatin pemetrexed; "within regimen"). Acceptable values are "within cancer" and "within regimen".

return_summary

Specifies whether a summary table for the cohort is returned. Default is FALSE. The `gtsummary` package is required to return a summary table.

Value

A list of data frames containing cancer diagnosis (`cohort_ca_dx`), cancer-directed regimen (`cohort_ca_drugs`) and next generation sequencing (`cohort_ngs`) information for patients that met the specified criteria. Optionally, if return_summary = TRUE, the list also includes summary tables for the number of records per dataset (`tbl_overall_summary`) as well as tables of key cancer diagnosis (`tbl_cohort`), cancer-directed regimen (`tbl_drugs`) and next generation sequencing (`tbl_ngs`) variables.

Author

Jessica Lavery

Examples

if(genieBPC:::check_synapse_login() == FALSE){
# Example 1 ----------------------------------
# Create a cohort of all patients with stage IV NSCLC adenocarcinoma and
# also return all of their corresponding cancer-directed drugs
nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")

create_analytic_cohort(cohort = "NSCLC",
  data_synapse = nsclc_2_1,
  stage_dx = "Stage IV",
  histology = "Adenocarcinoma")

# Example 2 ----------------------------------
# Create a cohort of all NSCLC patients who received Cisplatin,
# Pemetrexed Disodium or Cisplatin, Etoposide as their first drug regimen
# for their first index NSCLC
nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")

create_analytic_cohort(cohort = "NSCLC",
    data_synapse = nsclc_2_1,
    regimen_drugs = c("Cisplatin, Pemetrexed Disodium",
                      "Cisplatin, Etoposide"),
    regimen_order = 1,
    regimen_order_type = "within cancer")

# Example 3 ----------------------------------
# Create a cohort of all NSCLC patients who received Cisplatin, Pemetrexed
# Disodium at any time throughout the course of treatment for their
# cancer diagnosis,
# but in the event that the patient received the drug multiple times,
# only select the first time.
nsclc_2_1 <- pull_data_synapse("NSCLC", version = "2.1-consortium")

create_analytic_cohort(cohort = "NSCLC",
    data_synapse = nsclc_2_1,
    regimen_drugs = c("Cisplatin, Pemetrexed Disodium"),
    regimen_order = 1,
    regimen_order_type = "within regimen")
    }
#> Welcome, Michael Curry!
#> 
#> [1] "Great job! You are logged into Synapse and\n        can access the GENIE BPC data."
#> Welcome, Michael Curry!
#> 
#> Welcome, Michael Curry!
#> 
#> Welcome, Michael Curry!
#> 
#> $cohort_ca_dx
#> # A tibble: 181 x 152
#>    cohort record_id     institution ca_seq redcap_ca_index ca_cadx_int ca_dx_how
#>    <chr>  <chr>         <chr>        <int> <chr>                 <int> <chr>    
#>  1 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   18631 Pathology
#>  2 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   23329 Pathology
#>  3 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   23944 Pathology
#>  4 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   22538 Pathology
#>  5 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   25813 Pathology
#>  6 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   22189 Pathology
#>  7 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   22758 Pathology
#>  8 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   24688 Pathology
#>  9 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   24831 Pathology
#> 10 NSCLC  GENIE-DFCI-0~ DFCI             0 Yes                   22124 Pathology
#> # ... with 171 more rows, and 145 more variables: dob_next_ca_days <int>,
#> #   tt_first_index_ca <int>, naaccr_first_contact_int <int>, ca_age <int>,
#> #   age_dx <int>, naaccr_site_cd <chr>, ca_site <chr>, ca_d_site <chr>,
#> #   site_recode <chr>, ca_type <chr>, naaccr_histology_cd <int>,
#> #   ca_hist_adeno_squamous <chr>, naaccr_behavior_cd <int>,
#> #   naaccr_laterality_cd <chr>, naaccr_grade <int>, ca_grade_diff <chr>,
#> #   ca_grade <chr>, best_ajcc_stage_cd <chr>, ca_stage_iv <chr>, ...
#> 
#> $cohort_ca_drugs
#> # A tibble: 181 x 101
#>    cohort record_id  ca_seq institution regimen_number redcap_ca_index drugs_num
#>    <chr>  <chr>       <int> <chr>                <int> <chr>               <int>
#>  1 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  2 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  3 NSCLC  GENIE-DFC~      0 DFCI                     2 Yes                     2
#>  4 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  5 NSCLC  GENIE-DFC~      0 DFCI                     2 Yes                     2
#>  6 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  7 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  8 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#>  9 NSCLC  GENIE-DFC~      0 DFCI                     1 Yes                     2
#> 10 NSCLC  GENIE-DFC~      0 DFCI                     3 Yes                     2
#> # ... with 171 more rows, and 94 more variables: drugs_inst <chr>,
#> #   drugs_firstinst <chr>, drugs_ct_yn <chr>, drugs_dc_ynu <chr>,
#> #   regimen_drugs <chr>, drugs_drug_1 <chr>, drugs_drug_2 <chr>,
#> #   drugs_drug_3 <chr>, drugs_drug_4 <chr>, drugs_drug_5 <lgl>,
#> #   drugs_drug_oth_1 <chr>, drugs_drug_oth_2 <chr>, drugs_drug_oth_3 <lgl>,
#> #   drugs_drug_oth_4 <chr>, drugs_drug_oth_5 <lgl>, drugs_startdt_int_1 <int>,
#> #   drugs_startdt_int_2 <int>, drugs_startdt_int_3 <int>, ...
#> 
#> $cohort_ngs
#> # A tibble: 194 x 27
#>    cohort record_id         ca_seq institution cpt_number cpt_report_int
#>    <chr>  <chr>              <int> <chr>            <int>          <int>
#>  1 NSCLC  GENIE-DFCI-000263      0 DFCI                 1          18689
#>  2 NSCLC  GENIE-DFCI-000583      0 DFCI                 1          25932
#>  3 NSCLC  GENIE-DFCI-000611      0 DFCI                 1          24012
#>  4 NSCLC  GENIE-DFCI-000675      0 DFCI                 1          22604
#>  5 NSCLC  GENIE-DFCI-001319      0 DFCI                 1          25856
#>  6 NSCLC  GENIE-DFCI-002274      0 DFCI                 1          24349
#>  7 NSCLC  GENIE-DFCI-002433      0 DFCI                 1          24923
#>  8 NSCLC  GENIE-DFCI-003133      0 DFCI                 1          25920
#>  9 NSCLC  GENIE-DFCI-003307      0 DFCI                 1          25696
#> 10 NSCLC  GENIE-DFCI-003392      0 DFCI                 1          24054
#> # ... with 184 more rows, and 21 more variables: path_proc_number <int>,
#> #   path_rep_number <int>, cpt_genie_sample_id <chr>, cpt_oncotree_code <chr>,
#> #   cpt_sample_type <chr>, cpt_seq_assay_id <chr>, cpt_order_int <int>,
#> #   cpt_seq_date <int>, dob_cpt_report_days <int>, cpt_report_post_death <int>,
#> #   dx_cpt_rep_days <int>, dx_cpt_rep_mos <dbl>, dx_cpt_rep_yrs <dbl>,
#> #   dx_path_proc_cpt_days <int>, dx_path_proc_cpt_mos <dbl>,
#> #   dx_path_proc_cpt_yrs <dbl>, path_proc_cpt_rep_days <int>, ...
#>