Obtain Clinical Data Files for GENIE BPC Project

The `pull_data_synapse` function accesses the specified version of the clinical and genomic GENIE BPC data from Synapse and reads it into the R environment. Documentation corresponding to the clinical data files can also be found on Synapse in the Analytic Data Guide:

NSCLC v1.1-Consortium Analytic Data Guide

NSCLC v2.1-Consortium Analytic Data Guide

CRC v1.1-Consortium Analytic Data Guide

CRC v1.2-Consortium Analytic Data Guide

BrCa v1.1-Consortium Analytic Data Guide

pull_data_synapse(cohort, version)

Arguments

cohort
Vector or list specifying the cohort(s) of interest. Must be one of "NSCLC" (Non-Small Cell Lung Cancer) or "CRC" (Colorectal Cancer).

version
Vector or list specifying the version of the data. By default, the most recent version is pulled. Currently, only version 1.1 is available. When entering multiple cohorts, the order of the version numbers corresponds to the order that the cohorts are specified; the cohort and version number must be in the same order in order to pull the correct data. See examples below.

Value

Returns clinical and genomic data corresponding to the specified cohort(s). Data frames have the suffix indicating the cohort appended to their name, e.g. pt_char_NSCLC for the pt_char dataset of the NSCLC cohort.

Author

Mike Curry

Examples

# Example 1 ----------------------------------
# Pull the most recent non-small cell lung cancer data
# pull_data_synapse(cohort = "NSCLC", version = "2.1")

# Example 2 ----------------------------------
# Pull the most recent non-small cell lung cancer data and the most recent colorectal cancer data
# pull_data_synapse(cohort = c("NSCLC", "CRC"), version = c("2.1", "1.2"))

# Example 3 ----------------------------------
# Pull version 2.1 for non-small cell lung cancer and version 1.1 for colorectal cancer data
# pull_data_synapse(
#  cohort = c("NSCLC", "CRC"),
#  version = c("2.1", "1.1")
# )