Obtain clinical & genomic data files for GENIE BPC Project

The `pull_data_synapse` function accesses the specified version of the clinical and genomic GENIE BPC data from Synapse and reads it into the R environment. Documentation corresponding to the clinical data files can also be found on 'Synapse' in the Analytic Data Guide:

NSCLC v1.1-Consortium Analytic Data Guide

NSCLC v2.1-Consortium Analytic Data Guide

NSCLC v2.0-Public Analytic Data Guide

CRC v1.1-Consortium Analytic Data Guide

CRC v1.2-Consortium Analytic Data Guide

BrCa v1.1-Consortium Analytic Data Guide

Users must log in to 'Synapse' to access the data successfully. To set your 'Synapse' credentials during each session, call: `set_synapse_credentials(username = "your_username", password = "your_password")` To store authentication information in your environmental variables, add the following to your .Renviron file (tip: you can use usethis::edit_r_environ() to easily open/edit this file): `SYNAPSE_USERNAME = <your-username>` `SYNAPSE_PASSWORD = <your-password>` Alternatively, you can pass your username and password to each individual data pull function if preferred, although it is recommended that you manage your passwords outside of your scripts for security purposes.

Usage

pull_data_synapse(
  cohort = NULL,
  version = NULL,
  download_location = NULL,
  username = NULL,
  password = NULL
)

Arguments

cohort

Vector or list specifying the cohort(s) of interest. Must be one of "NSCLC" (Non-Small Cell Lung Cancer), "CRC" (Colorectal Cancer), or "BrCa" (Breast Cancer).

version

Vector specifying the version of the data. Must be one of the following: "v1.1-consortium", "v1.2-consortium", "v2.1-consortium", "v2.0-public". When entering multiple cohorts, the order of the version numbers corresponds to the order that the cohorts are specified; the cohort and version number must be in the same order in order to pull the correct data. See examples below.

download_location

if `NULL` (default), data will be returned as a list of dataframes with requested data as list items. Otherwise, specify a folder path to have data automatically downloaded there. When a path is specified, data are not read into the R environment.

username

'Synapse' username

password

'Synapse' password

Value

Returns a nested list of clinical and genomic data corresponding to the specified cohort(s).

Details

See the pull_data_synapse vignette for further documentation and examples.

Author

Karissa Whiting, Michael Curry

Examples

# Example 1 ----------------------------------
# Set up 'Synapse' credentials
set_synapse_credentials()
#> ✔ You are now connected to 'Synapse' as
#> bstgeniebpc@mskcc.org for this R session!

# Print available versions of the data
synapse_version(most_recent = TRUE)
#> # A tibble: 4 × 4
#>   cohort version         release_date versions_returned   
#>   <chr>  <chr>           <chr>        <chr>               
#> 1 BrCa   v1.1-consortium 2021-10      Most Recent Versions
#> 2 CRC    v1.2-consortium 2021-08      Most Recent Versions
#> 3 NSCLC  v2.1-consortium 2021-08      Most Recent Versions
#> 4 NSCLC  v2.0-public     2022-05      Most Recent Versions

# Pull version 2.0-public for non-small cell lung cancer
# and version 1.1-consortium for colorectal cancer data

 ex1 <- pull_data_synapse(
   cohort = c("NSCLC", "BrCa"),
   version = c("v2.0-public", "v1.1-consortium")
 )
#> ✔ pt_char has been imported for "BrCa_v1.1"
#> ✔ ca_dx_index has been imported for "BrCa_v1.1"
#> ✔ ca_dx_non_index has been imported for "BrCa_v1.1"
#> ✔ ca_drugs has been imported for "BrCa_v1.1"
#> ✔ prissmm_imaging has been imported for "BrCa_v1.1"
#> ✔ prissmm_pathology has been imported for "BrCa_v1.1"
#> ✔ prissmm_md has been imported for "BrCa_v1.1"
#> ✔ tumor_marker has been imported for "BrCa_v1.1"
#> ✔ cpt has been imported for "BrCa_v1.1"
#> ✔ fusions has been imported for "BrCa_v1.1"
#> ✔ cna has been imported for "BrCa_v1.1"
#> ✔ mutations_extended has been imported for "BrCa_v1.1"
#> ✔ pt_char has been imported for "NSCLC_v2.0"
#> ✔ ca_dx_index has been imported for "NSCLC_v2.0"
#> ✔ ca_dx_non_index has been imported for "NSCLC_v2.0"
#> ✔ ca_drugs has been imported for "NSCLC_v2.0"
#> ✔ prissmm_imaging has been imported for "NSCLC_v2.0"
#> ✔ prissmm_pathology has been imported for "NSCLC_v2.0"
#> ✔ prissmm_md has been imported for "NSCLC_v2.0"
#> ✔ cpt has been imported for "NSCLC_v2.0"
#> ✔ fusions has been imported for "NSCLC_v2.0"
#> ✔ cna has been imported for "NSCLC_v2.0"
#> ✔ mutations_extended has been imported for "NSCLC_v2.0"

 names(ex1)
#> [1] "BrCa_v1.1"  "NSCLC_v2.0"